update hash to int

+++ src/main/java/org/rev317/min/api/wrappers/SceneObject.java
@@ -34,7 +34,7 @@ public class SceneObject implements Locatable {
      * @return hash
      */
     public final int getHash() {
-        return accessor.getHash();
+        return (int) accessor.getHash();
     }
 
     /**
@@ -73,7 +73,7 @@ public class SceneObject implements Locatable {
      * @return region X
      */
     public final int getLocalRegionX() {
-        return accessor.getHash() & 0x7f;
+        return (int) accessor.getHash() & 0x7f;
     }
 
     /**
@@ -82,7 +82,7 @@ public class SceneObject implements Locatable {
      * @return region Y
      */
     public final int getLocalRegionY() {
-        return accessor.getHash() >> 7 & 0x7f;
+        return (int) accessor.getHash() >> 7 & 0x7f;
     }
 
     /**
@@ -91,7 +91,7 @@ public class SceneObject implements Locatable {
      * @return object id
      */
     public final int getId() {
-        return accessor.getHash() >> 14 & 0x7FFF;
+        return (int) accessor.getHash() >> 14 & 0x7FFF;
     }
 
     /**
